pcp-pmda-libvirt: rpm packaging changes
Turns out that on RH systems the required package changed name around
Fedora 27 from libvirt-python to python-libvirt (like the rest of
the world).

But the spec file had Requires: python-libvirt, which did not work
at all on older things like CentOS 7 or RHEL 7 where libvirt-python
is the only package option for the Requires: clause.

This is ugly and will need updating when the name change appears in
the regular RH release series (8?).