pmlogger_daily: fortify COMPRESS_DEFAULT with xz and options
Early versions of xz (e.g. on RHEL6) do not support xz --block-size.
Check for this and use xz with whatever options are available, overridden
by $PCP_COMPRESS if set. If xz is not installed and $PCP_COMPRESS is not
set, the code around line 1129 falls back to no compression.

This patch does not add any additional packaging dependencies on xz.
If it's not installed or not available on a particular platform, the
pmlogger_daily script will not do any compression.

QA group "logutil" passes (on F27 .. will also check on RHEL6)