32-bit only build configuration changes for OpenIndiana
It appears that Solaris, and hence OpenIndiana, has not engaged with
the Python cool aid to the extent of supporting 64-bit Python binaries.

This means the PCP Python shim has to be 32-bit, and the PCP libraries
need to be available as 32-bit ... we don't have the build infrastructure
to make 32-bit and 64-bit libraries in the one build (that magic was
unfortunately lost with IRIX decades ago), so it is 32-bit all the way
or nothing for Solaris and derivative SunOS travellers.

Changes touch configure, Makepkgs and qa/admin/myconfigure.