src/libpcp/src/auxconnect.c: add OpenIndiana fix to allow pmcd connection TCP
On OpenIndiana, __pmHostEntGetAddrInfo() can return a __pmAddrInfo
struct in which ai_socktype is 0 ... need to _not_ do the optimization
to skip entries with ai_socktype != SOCK_STREAM and fall through ...
to where things "just work".

Without this, any attempt to connect to pmcd from an OpenIndiana
platform to any external host, or itself via hostname or localhost,
would produce a ECONNREFUSED error because we never actually tried
to connect to anything.