assorted: be more careful with pid_t ... it may not be the same type as int
OpenIndiana found a bunch of compilation warnings associated with
code that assumed the type of pid_t and int are the same ... this
is not always true.

We do assume that a pid_t can be cast into an int without loss
of precision.  This is mostly for over the wire PDUs.

There is another case associated with pmdaRootProcessStart() et al
where the libpcp_pmda API has a pid argument declared as an int ...
this seems wrong, but fixing it would involve an API change that
I'm reluctant to embark upon.