Revert "32-bit only build configuration changes for OpenIndiana"
This reverts commit 869d8339723bc07c010dec33889cf8bc0e24e987.

Arrgh!

The Solaris PMDA is a /dev/kmem reader which means it has to match
the object code style of the kernel, namely 64-bit for 64-bit kernels.

So for Solaris (and hence OpenIndiana) we can have (a) a working
solaris PMDA, or (b) working Python apps, but not both.  This revert
reinstates option (a).