qa/admin/check-vm: some Perl modules cannot come from cpan for rpm builds
If a Perl module is (a) tested in configure.ac and (b) conditionally
used as a Requires pre-req in the spec file (both good), then the
module must NOT be installed from cpan.  Otherwise (a) passes and
(b) cause the rpm install to fail.

Initially this seems to apply to CentOS 6.9 (vm14) and ...

pmda-nutcracker
	YAML::XS::LibYAML and JSON
pmda-snmp
	Net::SNMP

Also update the n/a list for vm14 ... these Perl modules are not
available as RPMs.