qa: fix test 656 and common.qa handling of help text
The help text rebuild guard in the make_helptext function
was susceptible to short-circuit evaluation incorrectly in
the case of help file updates, but no help.dir rebuild.

This is fixed by making sure help.dir is newer than help,
and expected cifs metrics help text is updated in 656.out.