pmlogger_daily: add -p option for polling
and crontab glue to call this at 30mins past the hour every hour.

The intent here is that if the daily processing does not happen
for some reason at 00:10 (or whatever time the local crontab entry
may have been changed to), then we'd like to do the daily processing
as soon as possible ... if it has already been done, pmlogger_daily -p
exits.