libpcp_web: Add Cluster support to redis functionality.
We don't want to assume the redis topography being run by the local
admins.  Enabled cluster support by default, and changed from
redisContext's to redisSlots (collection of redisContexts, with a
single master node).  Existing TODO items include, a config file setup
(to avoid needing to manually point to a redis node), extension of the
globbing code to support the various testcases included in qa/1211, and
switch to asyc context types and callbacks.