src/pmlogger/utilproc.sh: add _save_prev_file()
Canonical and robust implementation when we want to save a prior
log file (or similar) as the same file name with a ".prev" suffix.

Trickiness here is mostly around existence of foo and/or foo.prev
at the time of the call, and various permissions wrinkles (mostly
from QA, rather than production use cases).