build/rpm/cherrypick: handle self-recursion in the dependency
pcp-pmda-activemq --requires PCP::Queue but also --provides PCP::Queue.

Before Queue became PCP::Queue this confused the hell out of me because
cherrypick refused to select pcp-pmda-activemq for installation because
the dependency on Queue was not satisfied, which caused pcp-collector
to fail to install and neither cpan nor yum helped find the Perl
Queue module (duh!).