qa/{098,100,101,103,104,105,106}: be more tolerant of pmcd state changes
Depending on which PMDAs are installed when the test is run,
their own internal state, and ... the phase of the moon, pmlogger
may see a pmcd state change, emit a diagnostic and force a <mark>
record into the archive.

If this happens, filter these artifacts out (they are not relevant
to the purpose of the test).