qa/515: tweak timing a little, rework "success" criteria
The timing sensitivity here is too large to be able to reliably
get a specific number, or even a range of numbers, of pmie alarms
being observed.

Relax the success criteria so that we pass if at least one of each
of the expected alarms is observed.  This still achieves the objective
of the test to exercise the rising and falling operators for pmie.