pcp2xlsx: switch from using xlsxwriter to openpyxl
For creating .xlsx files both xlsxwriter and openpyxl get the job done
but openpyxl is packaged for Fedora, xlsxwriter is not. openpyxl seems
also to be receiving more commits in upstream [1] and has more
stackoverflow traffic [2], so switch pcp2xlsx to use openpyxl instead
of xlsxwriter.

No functional change, the only slight downside is that the document
default font (Cambria) can't be changed but shouldn't matter. Passes
updated QA (need to see later if more tweaking is needed when updating
openpyxl version) and in general should work equally well as the
previous version. Can be finetuned further if we receive user feedback.

1)
https://github.com/jmcnamara/XlsxWriter
https://bitbucket.org/openpyxl/openpyxl/commits/all

2)
https://stackoverflow.com/tags/xlsxwriter
https://stackoverflow.com/tags/openpyxl