qa/{522,581,856,882,932,956,966,1000}: better control over systemd auto restart
We're seeing bad interactions between QA tests (that do not use
systemd) and systemd on some platforms, notably vm24 (openSUSE 42.2).
The symptoms are non-determinism because systemd may asynchronously
restart pmcd and/or pmlogger when the QA test was not expecting this.

There are more tests that need this treatment, but they have not
triggered failures ... this has to be viewed as good luck rather than
good engineering.  More changes will be coming.