qa/1343: small nit
systemd's autorestart should only be re-enabled if we (the QA test) can
get pmcd to start correctly ... so the _cleanup() ordering should be:

    _wait_for_pmcd
    ...
    _restore_auto_restart pmcd
not vice versa.  It does not matter until a rogue test leaves pmcd
in a crippled state (which should not happen, ha, ha).

I'll bet there are other tests that have this round the wrong way,
I just stumbled across it in qa/1343 while looking for something else.