pcp-dstat: initial version of dstat using PMAPI services
An implementation of the dstat utility that uses PCP - in
particular, building on pmconfig (pmrep) like the pcp2xxx
utilities.

The goals here are - a long-term supported dstat that uses
python2 or python3, has a configuration language, supports
PCP archives, remote hosts & local context mode, and is as
close to a drop-in replacement of the original dstat as we
can make it in terms of output.

Committing this draft version so that multiple people can
more easily hack on this together.  Code only, not built /
installed yet.  Reserves qa/1360, and man page to follow.
A fairly comprehensive TODO file accompanies this commit.