qa/1369: needs some systemd autorestart care for pmlogger
It is pmlogger that is exposed in this case ... the primary
pmlogger may terminate as a result of pmcd reconfiguration in
some earlier test (specifically the mmv metrics changed state).

It is much safter to disable pmlogger autorestart and stop pmlogger
at the start of the test.  And then in the _cleanup() routine start
pmlogger and re-enable autorestart.