zbxpcp: fixes, perf improvement for zabbix agent module
This commit tackles failures in qa/1067.  An error in the
return code handling was causing one set of issues, and a
lack of zabbix-agentd support for the config.d style conf
file setup was causing others (for some versions of the
zabbix-agentd).

Additionally, the startup code path is optimised somewhat
to reduce the number of PDUs we send to pmcd - this gives
an approx 3x speedup there, which helps for a short-lived
invocation situation like the QA test does.