src/pmdas/sample/src/sample.c: fix emerging gcc warning
gcc 8 in Fedora28 reports a warning for the following ...

sample.c: In function ‘init_tables’:
sample.c:1163:34: warning: argument to ‘sizeof’ in ‘strncpy’ call is the
same expression as the source; did you mean to use the size of the
destination? [-Wsizeof-pointer-memaccess]
     strncpy(_string, "13", sizeof("13"));

Turns out the warning is bogus, although the code looks kinda wrong
because sizeof("13") is 3 and this is the same as strlen("13")+1
(which would be a little more readable) and the 3rd argument should
be 3 here.

Made the same change in the earlier call to roundup() for local
consistency.

Thanks to Marko Myllynen for spotting this in the build logs.