Revert "pmlogrewrite: add -S (check and status) option"
This reverts commit 8d89a8843828c6a762d483e9ba2b139c3fc2419d.

On reflection, we don't need pmlogrewrite -S because

	pmlogrewrite -q -i arch

already does what's required, namely exit if there is not rewriting
required, plus the -q approach has the added benefit that the
rewriting will be done if any is required (which would be the most
likely action in a second pmlogrewrite invocation if -S returned
true).

Keeping the -S implementation in git, just in case anyone wants to
resurrect it at sometime in the future.