libpcp: pmOpenLog() support logging to existing stream (usually stderr)
PCP GH #365

This adds special handling to pmOpenLog() if the logname argument is "-".
Do not freopen the existing stream and do not rotate the log. The preamble
and atexit handler to write the footer on exit remain the same. There is
no effective change if the logname argument is not "-". See pmOpenLog(3)
for details.

All existing utilities and PMDAs using pmOpenLog(), i.e. supporting the
"-l logname" option can now write log messages to their existing stderr
stream by specifying -l-

Note that for PMDAs, if pmdaroot(1) is configured in pmcd.conf then it
will be the parent process and so any PMDA specifying -l- will write to
the root PMDA log stream, which may itself be stderr, in which case log
messages will appear in pmcd's log (which may also be stderr). If the
root PMDA is not in use, then pmcd will normally be the parent process.

Doc updates to pmOpenLog(3) and tested with qa/1255.

modified:   man/man3/pmopenlog.3
modified:   src/libpcp/src/util.c