pmdadm: fix segfault and memleak in dmstats
RH BZ #1569854

pmdadm dmstats was calling into libdevmapper with uninitialized
region and area ID arguments passed from the pw structure, resulting
in segfaults. Fix this by zeroing each pw structure immediately
after it's malloc'd. Once this was fixed, a memleak was identified
where only the last dm_stats handle was destroyed when traversing
all of the devices in a task - all but one handle thus leaked, so
this does not leak up if there is only one device.