qa/sniff4dodgey: (new) script looking for tests using selected dodgey features
Today, "dodgey" == "python" pretty much.  But the script does entertain
the notion of levels of dodgeyness.  At level 0 we have any test that
includes common.python, level 1 is not level 0 but the tests that use
python-related components that we expect to work across platforms,
level 5 is not levels 0-4 but tests that use a Python interpreter
directly, level 8 (default) is not levels 0-7 but the tests that use
use components with known cross-platform issues.