qa/group: update group tags, add missing python ones, introduce BAD
BAD is new.  If a test is tagged as BAD
- the tags also indicate which QA Farm machine it is failing on
- tests tagged BAD will NOT BE RUN on any machine in the QA Farm
  henceforth

Note the last point is a change in policy.  I'm sick of components
(Python in particular) that do not work, or are not available,
across platforms.

If you care about tests tagged "BAD" and want them run again in the
QA Farm you'll need to get in touch with me so you can understand
the issue and work on an acceptable resolution.