pmdabcc: fix tcplife with older/newer kernel/bcc
Upstream BCC tcplife (part of still unreleased 0.6.0) uses tracepoint
sock:inet_socket_set_state which seems to be the most suitable tp for
this. However, current Ubuntu LTS has only 4.15 which does not have
this tracepoint yet but instead it has tcp:tcp_set_state. So use
that one if available instead of falling back to fragile kbrobe based
approach. For this, we need different BPF files.

And if that's not enough, there's also a change in bcc upstream which
broke tcplife in upstream as well:

https://github.com/iovisor/bcc/commit/cb136c15470742699a250d8b4037098a9bd4fb17

The problem is that the old method fails with latest bcc while the new
method fails with older bcc. We need to detect which version works on
module compilation and select the version accordingly.

Tested on 4.15/4.16/4.17 with bcc-0.5.0/git, in all combinations
the expected variant is selected and the module is working (realistic
results are shown for processes).