pmdammv: resolve an instance domain setup problem
Several people (marko, lzap, tallpaul) have reported this
one, finally got to the bottom of it.  The symptoms are
"Unknown or illegal instance domain identifier" errors on
indom lookups, sometimes.  Root cause was a logic error
in pmdammv indom setup code incorrectly overwriting count
and offset local variables while parsing mappings.

To exercise the fix I've modernized qa/src/indom.c and
used it in new test qa/1422 to tickle the problem using
a canned MMV mapping which is known to expose it.

The original bug report (against Parfait) is this one:
https://github.com/performancecopilot/parfait/issues/53