qa/mk.logfarm, qa/637 & qa/641: (all new) pmlogger_daily setup infrastructure
The script sets up a forest of directories and PCP archives that mimic
a real setup as would be managed by pmlogger_check and pmlogger_daily ...
except the contents are deterministic.

We rewrite archives so that timestamps in the archive records are
"about" right for the datestamp from the name of the enclosing
archive, and make the access and modification times also "about"
right.