src/pmlogger/pmlogger_daily.sh: compress .meta files
Use the same -x and -X controls (or their environment variable
equivalents) as for the data volumes.

Small man page update also.

This is a preliminary checkin ... it is mostly working but I have
a couple of corner-cases that need some refinement (especially to
NOT compress the .meta file for the current pmlogger when -x 0
and/or -K is used), but I want to give this some air time and clear
the commit backlog.