pmParseTimeWindow(): fix double-free error seen from pmlogger_daily_report
Problem is deep inside libpcp, not pmlogger_daily_report, nor pmrep
that this script uses.

The management of the error/warning message buffer was broken in cases
where our internal parser failed but the fallback glib_get_date()
succeeds.

The man page was also misleading ... if the return value is 0 or -1,
then expect errMsg to point to a buffer that needs to be free()'d.