pmlogger: possible fix for trashed temporal index after infant mortality
I suspect that if pmlogger is killed off early enough, then we get
the broken data offset of 0 in the last termporal index entry.

This is almost impossible to test with QA directly, but the
check-archives script and the SaveLogs changes for
pmlogger_check/pmnewlog should enable us to see the problem go away
as this commit spreads over the QA Farm.