src/pmlogger/pmlogger_daily.sh: add some desperate debugging support
If the directory $PCP_LOG_DIR/pmlogger/SaveLogs exists, then the
log file from pmlogger_daily is saved there with a YYYYMMDD.HH.DD
component in the filename to reflect when pmlogger_daily was run.

This is no documented anywhere ... it may go away if it does not
prove useful.