pmdabcc: configurable failure modes for module errors
With this commit module failures configured to be fatal will now be
"aborting" in the log (in addition to earlier "Module NOT active") -
if a module fails to compile, compilation of a single module should
never take 30 seconds or more on a reasonable platform.  So, if the
"Ready to process requests" does not appear within a set timeframe
it could be used as a hint as well in QA that either the platform is
not suitable for QA (e.g., overloaded cloud/virt environments) or
something unexpected (e.g., BPF compilation or loading being stuck
which would be a severe kernel/tooling error) has occurred.

Fixes also a few import related pylint warnings seen with some pylint
versions.