libpcp: fork+exec and pipe+fork+exec changes for Windows
When this code was refactored for most platforms the Windows changes
(that are not built for other platforms) could not be checked.
Now we can check the code on Windows it was just broken in several
ways that made even the simplest operations fail ... and since almost
everything of interest (including the static binaries used in the
build) calls pmLoadNameSpace() and pmLoadNameSpace() runs pmcpp at
the end of a pipe, we were not making much progress in the build,
let alone running real code.

Changes here include getting the pipe file descriptors round the
right way (duh), more error checking, more -Dexec diagnostics,
not stripping leading drive and directory components of argv[0],
not creating a pipe in the exec-only case, ...