src/python/pmapi.c: Windows changes
1. don't include prototypes for methods defined in libpcp.h
2. if the platform headers do not define suseconds_t then there
   is no point adding a dictionary entry for PM_SIZEOF_SUSECONDS_T

Someone who understands this needs to check the 2. change ... I can see
no good reason for including PM_SIZEOF_SUSECONDS_T in the dictionary
ever ... the only places it or suseconds_t is used (in the entire
PCP src and qa trees) (a) in configure to set PM_SIZEOF_SUSECONDS_T
and (b) to add it into the Python dictionary in this source file.
It seems this whole box and dice could be nuked, n'est ce pas?