src/include/pcp/libpcp.h: important change for Windows
There is a botch in this header that is only exposed when the sizeof a
pointer or the sizeof a long is 64-bits and the sizeof a pointer is not
equal to the size of a long.

For out git-sdk-mingw compiler on Windows we have:
    #define HAVE_64BIT_PTR 1
    #define HAVE_32BIT_LONG 1

and ... kaboom when decoding a result PDU from a remote pmcd hits the
inconsistency between p_result.c in libpcp and the libpcp.h header.