libpcp (and pmlogger): change __pmNativePath() semantics
To support the Windows port where __pmNativePath() is something other
than the no-op it is on *nix platforms, we need the argument to be
*alloc'd and the function result needs to be used becuse on Windows
	new_path = __pmNativePath(path)
maybe (but not always) returns with new_path being completely different
(at least the leading part of it is different) to path.  So, path needs
to be *alloc'd, __pmNativePath may free(path) and will return a value
that is guaranteed to be *alloc'd.

For the *nix platforms, __pmNativePath unconditionally returns path
which is no change.

But to make it work for Windows the callers need to
- *alloc the input argument (not call with literal constants or a
  local char array)
- use the result of the __pmNativePath() call (and not assume the
  path rewriting happened in place in the input argument)
- free(path) if the first step forced an additional *alloc