src/pmcd/pmdaproc.sh: Windows changes
Because we're not yet at the bottom of the $PATH maze on Windows,
apps and scripts that live in $PCP_BINADM_DIR were not being run when
pmdaproc.sh was used.  We already had a sort of half-baked approach
to running some, but not all, of these as $PCP_BINADM_DIR/pmfoo rather
than pmfoo ... just make 'em all use the $PCP_BINADM_DIR/pmfoo style.

Also, for PMDA installation on Windows we need to play some really
ugly games with pathnames for DSO's and PMDA executables ... pwd(1)
produces a pathname that begins with a / but may not be a real path
to anything that you can execute or dlopen().  This is all down to
the half-baked chroot environment we're using (the comment in the
script sort of explains this, but as someone wiser than I once said
"you're not supposed to understand this").