assorted: be much more careful with (time_t *) casting
(applies to (const time_t *) casting as well)

The following is broken when time_t is 64-bits long and tv_sec
is 32-bits long (which is the default for 64-bit Windows):
	... pmCtime((time_t *)&foo.tv_sec, ...)

we need a local time_t, an explicit assignment and then pass the
time_t to pmCtime().

Note our PDU and archive formats are all safe as internally we use
32-bit values for "time in seconds" (not the platform's native time_t).
This will continue to work until 03:14:07 Jan 19 2038 by which time
I'll be almost 88 years old or dead, and in either case it will be
someone else's problem.