src/libpcp/src/exec.c: changes for __pmProcess*()
For __pmProcessCreate() change the arguments from the ambiguous and
confusing names infd and outfd to the clearer names fromChild and toChild
respectively.  I've wasted a lot of time being confused by this and the
alternate names make the situation clearer.

Also an important functional change in the Windows version of
__pmProcessPipe() ... because the "pipe" is built over a shared
handle for a Windows pipe, you cannot close the "end" you're not
going to use from the caller side ... this will close that "end"
on the spawned process side and nothing good will happen when
the child tries to read/write there.