qa: fix perfevent QA issues with f28 glibc/toolchain
It turns out that the gcc version in fedora 28 optimizes the malloc()
and subsequent memset() call in perf_event_create() to a call to calloc()
(which is not wrapped by the test harness). This is the reason why the
test_malloc_checking test fails.

The fix for this is to wrap calloc in the test harness.