src/pmdas/linux_proc/proc_pid.c: changes for /proc/N/environ masking
On at least one platform (vm21, Debian 7.11) we've been seeing qa/1222
failing because not all of the processes are in the instance domain.

After a lot of debugging, it appears that /proc/N/environ may be
inaccessible (EPERM) or appear empty ... either case was causing
a cascading error to bubble back up and invalidate most (but not all)
of the proc.psinfo instances.

Failing to get info from /proc/N/environ is no longer classified as
a major error that would invalidate other metrics that are available.