src/derived/iostat.conf: use rescale() to handle disk.dev.*_bytes units differences
On the *BSD platforms, the disk.dev.*_bytes metrics are in units of
bytes (for no really good reason) compared to kbytes on Linux platforms.

This breaks the derived metric expressions for the disk.dev.*_rqsz
metrics.

Using rescale() forces the metric values into a common unit and works
everywhere.