qa/1123: notrun on vm20
For reasons I can not discover, this test reliably hangs (and the
cron-driven QA never finishes) on vm20 (x86_64, Ubuntu 18.04) but
passes on other platforms including other x86_64 platforms running
the same Ubuntu version.

Simplest band aid is to not run it on vm20 ... it gets plenty of
exposure on other platforms.