qa/852: process 9268 is indeed exiting
In the fake linux root tarballs used by qa/852, $ROOT/proc/9268
is an empty directory, so trying to open $ROOT/proc/9268/status
(and similar files) fails. This is interpreted by the PMDA as a
process that is exiting. Unclear why that was not being reported
prior to recent changes in the proc_pmda - probably due to the
changes in some of the error handling in proc_pid.c. As far as
I can tell, it's correct, so just remake the qualified output.

modified:   qa/852.out