windows: fix a pmcd crash, a client-side issue and update services
Ensure we do not run off the end of the command argument array
(was causing a pmcd.exe crash in __pmServerStart).

Ensure the fdFlags array in libpcp connect logic is sized in a
way that is platform-independent.

Update the fallback paths for Win32 pcp-services.c to match the
sorts of locations we might find PCP services nowadays.