pminfo: change label reporting so metrics and instances are distinct
Its proving handy to be able to separately see the label components
for instances and metrics (in addition to indoms and instances, as
we've always had).  This is now showing up cases where cluster/item
labels are being set incorrectly - there seem to be some possible
quirks there in the Prometheus PMDA (qa/1221 shows it, fix to follow
later once diagnosed).

Tweak the BCC PMDA to allow labels to be set at indom level as well,
as this makes more sense for histograms where instance labels define
the buckets.  Previously we were using labels at the metric cluster
hierarchy level for this aspect - largely cosmetic, but we might as
well set this straight early on.