pmdaprometheus: instance labels were incorrectly returned as metric labels
Nathan's pminfo -l changes revealed pmdaprometheus was incorrectly
returning instance level labels as of the metric (item) level labels.
In addition, the "instname" label was not quite the same as the
indom external instance name, so fix this too.

qa/1221 and qa/1306 need an update for this change, coming next.