pminfo: revert labels reporting for instances behaviour
Chatting to Marko he found the change to not merge instance
labels all the way through the hierarchy non-obvious.  That
was done to expose some bugs, but end users don't expect it
so change it back.  We can still see those bugs due to the
metric-level and indom-level label reporting (which remains
intact), so this change does seem like the best option.