pmdapostgresql: intialize pmdaCache files to be group writeable
postgres user needs to be able to write to the pmdaCache files
(one for each indom) but can't create them itself. So we create
them in the Install script (if they don't already exist) and make
them group writeable (mode 664) for the postgres user's gid.

modified:   src/pmdas/postgresql/Install