pmproxy: initial refactored code for RESP proxying
Further strides toward proxying Redis protocol in pmproxy, adding
in more (refactored) work from Prajwal's GSoC project.

This adds a dictionary onto the Redis slots structure in libpcp_web
maintaining a mapping of commands to first-key-position such that a
Redis server selection can be made based on key as described in the
cluster specification, and as used in the PCP time series code too.