pmdabcc: allow modules to bypass pmdaCache
In a rare case a module needs to bypass pmdaCache it would now be
possible. Since this would require supporting code from modules
this cannot be enabled currently with any included modules so we
leave this intentionally undocumented (and low-level details like
pmdaCache should not be even mentioned for casual users).