pmdabcc: support call latency reporting in ucalls module
Similar in upstream except that no syscall count support which is
already provided by the syscount module.

Since the module might be running longer than standalone tools we
use usec not nsec - some accuracy is lost but unlikely to be that
relevant. In theory usec might also overflow at some point but ms
instead of usec would perhaps lose too much accuracy. Need to see
how this work in real life and if needed switch to ms if must.