pmdabcc: compat / QA tweaks
Add new helper to return BCC version in pcpbcc.python. Use it.
Use it in ustat.python as well and add compat for bcc-0.5.0.
Add similar standalone-script to allow skipping QA which will
fail in case of BCC issues. Use it in QA 1170. Exit hard from
tcptop module if incompabile BCC version in use (and which
can't be worked around). Tweak QA 1178 for older kernels.