build: add a 'make check' target, available everywhere
This adds a default 'check' target to all components
of the build (via buildrules) which can be optionally
extended in the makefile for each tool/library.

This makes it easier to invoke each of the local unit
and pylint tests from one place (descend from $TOPDIR)
so that %check in RPM packaging can perform this task,
as can Travis; each of those will arrive in subsequent
commits.

configure.ac checking for pylint and cppcheck is added.
The generic .pylintrc config (from the BCC efforts) is
now made available to all python tools via new $PYLINT
macro.