pmdads389: make perl-LDAP a softdep on some platforms
The ds389 PMDA uses perl-LDAP but this isn't available on some RPM
based platforms. After some discussion, we decided to make this a
softdep in the RPM specs with a runtime check in the PMDA to avoid
confusing installation failures (especially when installing pcp-collector).

Longer term, this PMDA should probably be re-written in python.