pcp-dstat: fix printing of empty lines with no-values-available
Recent fixes to dstat to handle no-values correctly have exposed
an odd-looking behaviour where full rows are printed without any
values - this is common for the first sample when all plugins we
are displaying use counter metrics.

Refactor the cprint() code (and helpers) to help detecting this
situation - all are now member functions and clear the novalues
field as needed - and tweak the reporting logic to drop lines
with no values at all.