install-sh: fix for *BSD strangeness
Some (maybe all) of my BSD builds are failing in man/html/images
where the install-sh invocation is dying in assorted strange ways
e.g. not recognizing the command line options and trying to
    cp -o .../-o
which is not going to work.

The use of getopts in install-sh is a little odd ... the $* at the end
is not needed and without this things work on *BSD.

There are no apparent recent changes in install-sh that would explain
this and none of the *BSD systems have been recently upgraded.

All I can think of is that some image files have been added as targets
in man/html/images/GNUmakefile and either the length of the resulting
command line or some wierd characters in the file names is confusing
getopts.