pmdabcc: add high-resolution pid tracking alternative to tcpperpid
The BCC PMDA tcptop module uses a proven "high-resolution" BPF program
to trace TCP connection transfer amounts. However, the downside is
it comes with (much) higher overhead [1] than the approach so far used
by tcpperpid (which is based on tracepoints on socket state changes).
So using the high-res variant would not be acceptable in certain envs.

Allow using the high-res BPF program alternative which is already used
by tcptop optionally, the default remains as the low-overhead variant
which is likely to be well adequate in many if not even most cases.

See the note at the top of upstream tcptop about description of the
possible overhead:

https://github.com/iovisor/bcc/blob/bfbf17ea425c087b110854ac881c351c6802f1fe/tools/tcptop.py#L12