pmdapostgresql: exercise new cursors to make sure they're valid
On some platforms, after bouncing the postgresql server, existing
cursors think they're still valid. So execute a simple query to
exercise a new cursor which throws an exception if it's invalid;
triggers a db reconnect if so. Was noticed with qa/1362 on f28
with postgresql v10.2 but not on f28 with v9.x.

modified:   src/pmdas/postgresql/pmdapostgresql.python