pmdapostgresql: minor tweak to work with postgresql v8.x
Handle KeyError exception when key column name is not present
in a known table. e.g. in postgresql v8.4.20 on RHEL6.9, the
'pid' column in the pg_stat_activity table is called 'procpid'.

Use alternate form of psycopg2.connect(), where the dbname,
user and password are all passed in a single string - this
works on v8, v9 and v10.

modified:   src/pmdas/postgresql/pmdapostgresql.python