src/pmns/Rebuild: conditionally omit some root_foo PMNS files
This fixes a problem that has existed for years and I've previously
been mis-diagnosing as QA test sloppiness.  The symtom is a subtree
in the PMNS (usually jbd2 or kvm or mmv) when the corresponding PMDA
is NOT configured in pmcd.conf.

The root cause ended up being the Rebuild script (which is always
called for the first pmcd start after an upgrade).  More careful
checking here will only include a root_foo file in the PMNS if the
associated foo PMDA is configured in pmcd.conf.

There is (unused today) support for one-level mapping if the metrics
in the root_foo PMNS file are instantiated by the bar (not foo) PMDA,
and also support for some root_bogus files to be unconditionally
ignored (to preserve a historical anomaly).

Also added a -o option to preserve the old behaviour, as a couple of
QA tests rely on this.