qa/1267: remove problematic diff use
On vm14 the context around the diff reporting (and the line numbers)
are not quite right.  Replace diff with _pick() (new) to report just
the lines of interest from the pmdumptext output.

Also a little bit of pretty messing with the output and how much is
syphoned off to $seq.full.