qa/1349: remove problematic diff use
(same treatment as qa/1267)

On vm14 the context around the diff reporting (and the line numbers)
are not quite right.  Replace diff with _pick*() (new, one per test
case) to report just the lines of interest from the pmdumptext output.

Also a little bit of pretty messing with the output and how much is
syphoned off to $seq.full.