qa/admin/check-vm: massive FreeBSD changes
Turns out that FreeBSD and OpenIndiana both use pkg(1) for packaging,
but these are entirely different commands with different repo naming
schemes, which explains why this was so confused as I moved between
various VMs. So split pkg? lines into F_pkg? (for FreeBSD) and
S_pkg? (for Solaris aka OpenIndiana), and add goo to conditionally
map as appropriate.