pmlogrewrite(1): Allow instance names to be used to specify instances labels.
In addition to the currently supported instance number, an instance
name in double quotes may now also be used to identify the individual
instance with which labels to be rewritten are associated.

The normal rule of matching instance names up to the first whitespace
applies.

QA and man page updates are included in these changes.