pmlogger_daily: change workflow to cull early
Change the workflow to move culling of old files earlier in the
pipeline.  So if there is a problem with log rewriting or merging that
is unattended for a log time, we'll eventually cull the offending
archives rather than having the rewriting or mering failure block
the culling and (in extreme cases) lead to full flesystems

Also add qa/686 (new) to check all of this out.