qa: tweak qa/1044 for travis-ci automated testing
qa/1044 has been failing in the travis-ci qa sanity test.
This seems to be because the first time pmieconf is run on
a new rules directory, it prints a message (and that message
is unexepected in the qualified output of the qa test). Run
it again and it always passes.

In the travis-ci testing, a new container image is always built
and a the just-built pcp packages are installed into that container.
So we're hitting the first-run issue with that test.

To fix this, the qa test now runs pmieconf and discards the output
before running the rest of the tests.

Also updated the travis-ci REDME with additional stuff about how
to manually run what travis-ci tests.

modified:   build/containers/travis-ci/README
modified:   qa/1044