build: add --check option to Makepkgs script
Issue make check as part of travis CI from within the build
environment.  This allows us to use a more recent toolchain
available from within the (fedora) container; including the
cppcheck and pylint from earlier (now reenabled).

As part of resolving this, I've fixed the mafefiles for the
pidstat and mpstat utilities which were doing unholy things
with symlinks committed to git (these links are now LDIRT).
This happens to fix issues in the Windows build too.