libpcp_web, pmproxy: refactor time series load code for sharing
Major refactor of the time series load code to allow both the
manual loading mechanism and automatic archive discovery to be
able to share internal code and data structures.

This also includes a significant update to the way string maps
are handled throughout libpcp_web, simplifying their use using
identifiers that can be calculated client-side.  This removes
several classes of round-trip on string evaluation.

A number of reference counting issues in querying are resolved
here as well.  Small updates to match Redis error handling now
in use in the released v5 tree.