pmchart(1): Fix incorrect use of std::vector::resize().
This method takes a number of elements but was being passed a
size in bytes. This had the effect of adding extra
(uninitialized) members to the
vector which is used to hold graph data samples.

There was no ill effect when going forward since new samples are
added to the beginning of the vector. However, when going backward,
new samples are added to the end of the vector, well beyond the
range of the samples used to create the graph. The result was
random graph displays when going backward.