configure.ac: change logic for fmt_int64 on the "hardway" path
We used to check for d, ld and then lld.  The check is very wonky
donkey and failed miserably on OpenBSD.  Reversing the order of the
tests works for OpenBSD, but I'll have to wait to see if this ends
up being wrong for other platforms.  Fortunately most platforms do
not need the "hardway" logic, so they should see no change.